requirements.txt
flask
tensorflow
numpy
pandas
scikit-learn
pillow
gunicorn